sum fn needs rework

+++ src/main.rs
@@ -29,10 +29,10 @@ fn main() -> Result<()> {
    
     
     let cases: Cases = Cases::new();
-    let cities : Vec<u16>= cases.l40;
+    let cities : Vec<usize>= cases.l40;
     let mut sa: SimAnn = SimAnn::new(cities.len().try_into().unwrap(), &cities);
     sa.prepare();
-       // sa.fill_distances();
+    
     
     let mut sol : Solution = Solution::new(0.002, 8.0, 0.95, &cities, 2000);
     sol.threshold_acceptance();
